SDI-266 Refactor movements and renamed tests for consistency
