+++ src/main/java/uk/gov/justice/hmpps/prison/service/receiveandtransfer/ExternalMovementTransferService.kt
@@ -4,6 +4,7 @@ import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import uk.gov.justice.hmpps.prison.api.model.RequestToTransferIn
 import uk.gov.justice.hmpps.prison.repository.jpa.model.AgencyLocation
+import uk.gov.justice.hmpps.prison.repository.jpa.model.City
 import uk.gov.justice.hmpps.prison.repository.jpa.model.CourtEvent
 import uk.gov.justice.hmpps.prison.repository.jpa.model.ExternalMovement
 import uk.gov.justice.hmpps.prison.repository.jpa.model.MovementDirection
@@ -36,34 +37,14 @@ class ExternalMovementTransferService(
     val receiveDateTime = getReceiveDateTime(request.receiveTime, booking).getOrThrow()
     val movementType = getAdmissionMovementType().getOrThrow()
 
-    return with(booking) {
-      setPreviousMovementsToInactive().also { entityManager.flush() }
-      addExternalMovement(
-        ExternalMovement(
-          /* offenderBooking = */ booking,
-          /* movementSequence = */ null,
-          /* movementDate = */ receiveDateTime.toLocalDate(),
-          /* reportingDate = */ null,
-          /* movementTime = */ receiveDateTime,
-          /* eventId = */ null,
-          /* parentEventId = */ null,
-          /* arrestAgencyLocation = */ null,
-          /* fromAgency = */ lastMovement.fromAgency,
-          /* toAgency = */ lastMovement.toAgency,
-          /* active = */ true,
-          /* escortText = */ null,
-          /* escortCode = */ null,
-          /* commentText = */ request.commentText,
-          /* toCity = */ null,
-          /* fromCity = */ null,
-          /* movementReason = */ movementReason,
-          /* movementDirection = */ MovementDirection.IN,
-          /* movementType = */ movementType,
-          /* toAddressId = */ null,
-          /* fromAddressId = */ null,
-        )
-      )
-    }
+    return booking.addExternalMovementIn(
+      movementDateTime = receiveDateTime,
+      fromLocation = lastMovement.fromAgency,
+      prison = lastMovement.toAgency,
+      commentText = request.commentText,
+      movementReason = movementReason,
+      movementType = movementType,
+    )
   }
 
   fun updateMovementsForCourtTransferToDifferentPrison(
@@ -77,34 +58,14 @@ class ExternalMovementTransferService(
     val receiveDateTime = getReceiveDateTime(movementDateTime, booking).getOrThrow()
     val movementType = getAdmissionMovementType().getOrThrow()
 
-    return with(booking) {
-      setPreviousMovementsToInactive().also { entityManager.flush() }
-      addExternalMovement(
-        ExternalMovement(
-          /* offenderBooking = */ booking,
-          /* movementSequence = */ null,
-          /* movementDate = */ receiveDateTime.toLocalDate(),
-          /* reportingDate = */ null,
-          /* movementTime = */ receiveDateTime,
-          /* eventId = */ null,
-          /* parentEventId = */ null,
-          /* arrestAgencyLocation = */ null,
-          /* fromAgency = */ lastMovement.fromAgency,
-          /* toAgency = */ toAgency, // the passed in agency in the request is just for validation
-          /* active = */ true,
-          /* escortText = */ null,
-          /* escortCode = */ null,
-          /* commentText = */ commentText,
-          /* toCity = */ null,
-          /* fromCity = */ null,
-          /* movementReason = */ movementReason,
-          /* movementDirection = */ MovementDirection.IN,
-          /* movementType = */ movementType,
-          /* toAddressId = */ null,
-          /* fromAddressId = */ null,
-        )
-      )
-    }
+    return booking.addExternalMovementIn(
+      movementDateTime = receiveDateTime,
+      fromLocation = lastMovement.fromAgency,
+      prison = toAgency,
+      commentText = commentText,
+      movementReason = movementReason,
+      movementType = movementType,
+    )
   }
 
   fun updateMovementsForCourtTransferToSamePrison(
@@ -119,35 +80,18 @@ class ExternalMovementTransferService(
     val receiveDateTime = getReceiveDateTime(movementDateTime, booking).getOrThrow()
     val movementType = getCourtMovementType().getOrThrow() // not an admission as only returning from court
 
-    return with(booking) {
-      setPreviousMovementsToInactive().also { entityManager.flush() }
-      addExternalMovement(
-        ExternalMovement(
-          /* offenderBooking = */ booking,
-          /* movementSequence = */ null,
-          /* movementDate = */ receiveDateTime.toLocalDate(),
-          /* reportingDate = */ null,
-          /* movementTime = */ receiveDateTime,
-          /* eventId = */ courtEvent?.id,
-          /* parentEventId = */ courtEvent?.parentCourtEventId,
-          /* arrestAgencyLocation = */ null,
-          /* fromAgency = */ lastMovement.toAgency,
-          /* toAgency = */ lastMovement.fromAgency,
-          /* active = */ true,
-          /* escortText = */ null,
-          /* escortCode = */ null,
-          /* commentText = */ commentText,
-          /* toCity = */ null,
-          /* fromCity = */ null,
-          /* movementReason = */ movementReason,
-          /* movementDirection = */ MovementDirection.IN,
-          /* movementType = */ movementType,
-          /* toAddressId = */ null,
-          /* fromAddressId = */ null,
-        )
-      )
-    }
+    return booking.addExternalMovementIn(
+      movementDateTime = receiveDateTime,
+      eventId = courtEvent?.id,
+      parentEventId = courtEvent?.parentCourtEventId,
+      fromLocation = lastMovement.toAgency,
+      prison = lastMovement.fromAgency,
+      commentText = commentText,
+      movementReason = movementReason,
+      movementType = movementType,
+    )
   }
+
   fun updateMovementsForTransferInAfterTemporaryAbsenceToSamePrison(
     movementReasonCode: String?,
     movementDateTime: LocalDateTime?,
@@ -160,34 +104,19 @@ class ExternalMovementTransferService(
     val receiveDateTime = getReceiveDateTime(movementDateTime, booking).getOrThrow()
     val movementType = getTAPMovementType().getOrThrow()
 
-    return with(booking) {
-      setPreviousMovementsToInactive().also { entityManager.flush() }
-      addExternalMovement(
-        ExternalMovement(
-          /* offenderBooking = */ booking,
-          /* movementSequence = */ null,
-          /* movementDate = */ receiveDateTime.toLocalDate(),
-          /* reportingDate = */ null,
-          /* movementTime = */ receiveDateTime,
-          /* eventId = */ scheduleEvent?.id,
-          /* parentEventId = */ scheduleEvent?.parentEventId,
-          /* arrestAgencyLocation = */ null,
-          /* fromAgency = */ lastMovement.toAgency,
-          /* toAgency = */ lastMovement.fromAgency,
-          /* active = */ true,
-          /* escortText = */ null,
-          /* escortCode = */ lastMovement.escortCode,
-          /* commentText = */ commentText,
-          /* toCity = */ null,
-          /* fromCity = */ lastMovement.toCity,
-          /* movementReason = */ movementReason,
-          /* movementDirection = */ MovementDirection.IN,
-          /* movementType = */ movementType,
-          /* toAddressId = */ null,
-          /* fromAddressId = */ lastMovement.toAddressId,
-        )
-      )
-    }
+    return booking.addExternalMovementIn(
+      movementDateTime = receiveDateTime,
+      eventId = scheduleEvent?.id,
+      parentEventId = scheduleEvent?.parentEventId,
+      fromLocation = lastMovement.toAgency,
+      prison = lastMovement.fromAgency,
+      escortCode = lastMovement.escortCode,
+      commentText = commentText,
+      fromCity = lastMovement.toCity,
+      movementReason = movementReason,
+      movementType = movementType,
+      fromAddressId = lastMovement.toAddressId,
+    )
   }
 
   fun updateMovementsForTransferInAfterTemporaryAbsenceToDifferentPrison(
@@ -201,34 +130,14 @@ class ExternalMovementTransferService(
     val receiveDateTime = getReceiveDateTime(movementDateTime, booking).getOrThrow()
     val movementType = getAdmissionMovementType().getOrThrow()
 
-    return with(booking) {
-      setPreviousMovementsToInactive().also { entityManager.flush() }
-      addExternalMovement(
-        ExternalMovement(
-          /* offenderBooking = */ booking,
-          /* movementSequence = */ null,
-          /* movementDate = */ receiveDateTime.toLocalDate(),
-          /* reportingDate = */ null,
-          /* movementTime = */ receiveDateTime,
-          /* eventId = */ null,
-          /* parentEventId = */ null,
-          /* arrestAgencyLocation = */ null,
-          /* fromAgency = */ lastMovement.fromAgency,
-          /* toAgency = */ toAgency, // the passed in agency in the request is just for validation
-          /* active = */ true,
-          /* escortText = */ null,
-          /* escortCode = */ null,
-          /* commentText = */ commentText,
-          /* toCity = */ null,
-          /* fromCity = */ null,
-          /* movementReason = */ movementReason,
-          /* movementDirection = */ MovementDirection.IN,
-          /* movementType = */ movementType,
-          /* toAddressId = */ null,
-          /* fromAddressId = */ null,
-        )
-      )
-    }
+    return booking.addExternalMovementIn(
+      movementDateTime = receiveDateTime,
+      fromLocation = lastMovement.fromAgency,
+      prison = toAgency,
+      commentText = commentText,
+      movementReason = movementReason,
+      movementType = movementType,
+    )
   }
 
   fun updateMovementsForNewOrRecalledBooking(
@@ -242,34 +151,14 @@ class ExternalMovementTransferService(
     val movementReason = getMovementReason(movementReasonCode).getOrThrow()
     val movementType = getAdmissionMovementType().getOrThrow()
 
-    return with(booking) {
-      setPreviousMovementsToInactive().also { entityManager.flush() }
-      addExternalMovement(
-        ExternalMovement(
-          /* offenderBooking = */ booking,
-          /* movementSequence = */ null,
-          /* movementDate = */ receiveDateTime.toLocalDate(),
-          /* reportingDate = */ null,
-          /* movementTime = */ receiveDateTime,
-          /* eventId = */ null,
-          /* parentEventId = */ null,
-          /* arrestAgencyLocation = */ null,
-          /* fromAgency = */ fromLocation,
-          /* toAgency = */ prison, // the passed in agency in the request is just for validation
-          /* active = */ true,
-          /* escortText = */ null,
-          /* escortCode = */ null,
-          /* commentText = */ commentText,
-          /* toCity = */ null,
-          /* fromCity = */ null,
-          /* movementReason = */ movementReason,
-          /* movementDirection = */ MovementDirection.IN,
-          /* movementType = */ movementType,
-          /* toAddressId = */ null,
-          /* fromAddressId = */ null,
-        )
-      )
-    }
+    return booking.addExternalMovementIn(
+      movementDateTime = receiveDateTime,
+      fromLocation = fromLocation,
+      prison = prison,
+      commentText = commentText,
+      movementReason = movementReason,
+      movementType = movementType,
+    )
   }
 
   private fun getAdmissionMovementType(): Result<MovementType> =
@@ -281,6 +170,7 @@ class ExternalMovementTransferService(
     movementTypeRepository.findByIdOrNull(MovementType.CRT)?.let { success(it) } ?: failure(
       EntityNotFoundException.withMessage("No ${MovementType.CRT} movement type found")
     )
+
   private fun getTAPMovementType(): Result<MovementType> =
     movementTypeRepository.findByIdOrNull(MovementType.TAP)?.let { success(it) } ?: failure(
       EntityNotFoundException.withMessage("No ${MovementType.TAP} movement type found")
@@ -314,4 +204,44 @@ class ExternalMovementTransferService(
     externalMovementRepository.findAllByOffenderBooking_BookingIdAndActive(this.bookingId, true).any {
       movementTime.isBefore(it.movementTime)
     }
+
+  private fun OffenderBooking.addExternalMovementIn(
+    movementDateTime: LocalDateTime,
+    eventId: Long? = null,
+    parentEventId: Long? = null,
+    fromLocation: AgencyLocation? = null,
+    prison: AgencyLocation,
+    escortCode: String? = null,
+    commentText: String? = null,
+    fromCity: City? = null,
+    movementReason: MovementReason,
+    movementType: MovementType,
+    fromAddressId: Long? = null,
+  ): ExternalMovement = this.setPreviousMovementsToInactive().also { entityManager.flush() }.let {
+    this.addExternalMovement(
+      ExternalMovement(
+        /* offenderBooking = */ this,
+        /* movementSequence = */ null,
+        /* movementDate = */ movementDateTime.toLocalDate(),
+        /* reportingDate = */ null,
+        /* movementTime = */ movementDateTime,
+        /* eventId = */ eventId,
+        /* parentEventId = */ parentEventId,
+        /* arrestAgencyLocation = */ null,
+        /* fromAgency = */ fromLocation,
+        /* toAgency = */ prison, // the passed in agency in the request is just for validation
+        /* active = */ true,
+        /* escortText = */ null,
+        /* escortCode = */ escortCode,
+        /* commentText = */ commentText,
+        /* toCity = */ null,
+        /* fromCity = */ fromCity,
+        /* movementReason = */ movementReason,
+        /* movementDirection = */ MovementDirection.IN,
+        /* movementType = */ movementType,
+        /* toAddressId = */ null,
+        /* fromAddressId = */ fromAddressId,
+      )
+    )
+  }
 }
