+++ src/test/java/uk/gov/justice/hmpps/prison/api/resource/impl/OffenderResourceIntTest_newOffender.kt
@@ -16,7 +16,7 @@ import java.time.LocalDate
 import java.time.temporal.ChronoUnit
 
 @WithMockUser
-class OffenderResourceNewOffenderTest : ResourceTest() {
+class OffenderResourceIntTest_newOffender : ResourceTest() {
 
   @Nested
   @DisplayName("POST /offenders")
