--- src/test/resources/uk/gov/justice/hmpps/prison/api/resource/impl/new_booking_prisoner.json
@@ -1,34 +1,0 @@
-{
-  "offenderNo": "Z0022ZZ",
-  "offenderId": -1022,
-  "rootOffenderId": -1022,
-  "firstName": "HARVEY",
-  "lastName": "HARRISON",
-  "dateOfBirth": "1962-01-01",
-  "activeFlag": true,
-  "agencyId": "SYI",
-  "assignedLivingUnitId": -202,
-  "assignedLivingUnit": {
-    "agencyId": "SYI",
-    "locationId": -202,
-    "description": "A-1-1",
-    "agencyName": "Shrewsbury"
-  },
-  "physicalAttributes": {
-    "gender": "Male",
-    "raceCode": "W1",
-    "ethnicity": "White: British",
-    "sexCode": "M"
-  },
-  "profileInformation": [
-    {
-      "type": "YOUTH",
-      "question": "Youth Offender?",
-      "resultValue": "Yes"
-    }
-  ],
-  "inOutStatus": "IN",
-  "status": "ACTIVE IN",
-  "lastMovementTypeCode": "ADM",
-  "lastMovementReasonCode": "24"
-}
